Handles more types of forms, better infinite loop detection
This handles more types of forms, such as the data-sharing dialogue that
pops up every now and then when logging into LADOK.

This also improves the loop detection code so that we can detect bigger
cycles than just the previous URL being the same. Also we don't just
look at the URL, as that will be the same for the data-sharing dialogue,
but also look at the inputs in the form (which will differ).